Set Virtual4DSG renderer

+++ libDataLoader/include/dataLoader/CameraParameters.h
@@ -2,11 +2,14 @@
 #define _H_PSLAM_CAMERA_PARAMETERS
 namespace PSLAM {
     struct CameraParameters {
-        float cx, cy;
-        float fx, fy;
+        // float cx, cy;
+        // float fx, fy;
+        double cx, cy;
+        double fx, fy;
         unsigned short width, height;
         float scale = 1.f;
-        void Set(unsigned int width, unsigned int height, float fx, float fy, float cx, float cy, float scale = 1.f) {
+        // void Set(unsigned int width, unsigned int height, float fx, float fy, float cx, float cy, float scale = 1.f) {
+        void Set(unsigned int width, unsigned int height, double fx, double fy, double cx, double cy, float scale = 1.f) {
             this->width = width;
             this->height = height;
             this->fx = fx;
